Add parser to clean up code
