+++ src/shared/Parser.ts
@@ -1,0 +1,6 @@
+import { Page } from 'puppeteer';
+
+export interface Parser<Type> {
+  parse: (page: Page) => Promise<void>;
+  getObject: () => Type;
+}
